requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-core>=0.1.0
langsmith>=0.0.75
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
msal>=1.22.0
pytz>=2023.3
flask>=2.2.3
supabase>=2.0.0
